+++ src/api/clients/TMDBClient.ts
@@ -1,7 +1,8 @@
 import { Movie } from '../../domain/movie.type';
 import { Client } from '../Client';
 import { MoviesAPI, MoviesResponse, PaginatedParams } from '../interfaces/MoviesAPI';
-import { TMDBAPISpec, TMDBClientProps, TMDBLanguage, TMDBMovie } from './TMDBClient.type';
+import { TMDBAPISpec } from './TMDBClient.api';
+import { TMDBClientProps, TMDBLanguage, TMDBMovie } from './TMDBClient.type';
 
 export class TMDBClient extends Client<TMDBAPISpec> implements MoviesAPI {
   apiKey: string;
